% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{doLambda-package}
\alias{doLambda-package}
\title{A Lambda parallel back end for foreach.}
\description{
The doLambda package implements a backend for using AWS lambda as
the backend, with S3 serving as the job queue.
}
\seealso{
\code{\link{registerDoLambda}}
}
