% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doLambda.R
\name{registerDoLambda}
\alias{registerDoLambda}
\title{registerDoLambda}
\usage{
registerDoLambda(bucket, key = NULL, secret = NULL, region = NULL)
}
\arguments{
\item{bucket}{name of an S3 bucket you own that will hold the job queue}

\item{key}{AWS Access Key.}

\item{secret}{AWS Secret Key.}

\item{region}{AWS Default Region.}
}
\value{
nothing. Called for side effect of registering doParallel backend.
}
\description{
Tell doPar to use this backend.
}
\note{
If credentials are not provided (key, secret, and region), asw,signature 
will be used to locate them by searching in the following order: environment 
variables (`AWS_ACCESS_KEY`, `AWS_SECRET_ACCESS_KEY`, and `AWS_DEFAULT_REGION`),
./.aws/credentials, and finally ~/.aws/credentials. If no credentials can be 
found, the function will throw an error.
}
