% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{lambdaWorker}
\alias{lambdaWorker}
\title{lambdaWorker}
\usage{
lambdaWorker()
}
\value{
nothing. Called for side effect of doing a unit of work.
}
\description{
Retrieve job and execute it.
}
\note{
This function is automatically called (by the boostrap script) when the 
lambda job is triggered. It typically is never used on the client end.
}
